Refactor: dir structure
