+++ packages/data-broker/src/__tests__/services/dhis/pullers/DeprecatedEventsPuller.test.js
@@ -6,7 +6,7 @@
 import { createModelsStub, stubDhisApi } from '../DhisService.stubs';
 import { DATA_GROUPS } from '../DhisService.fixtures';
 import { DeprecatedEventsPuller } from '../../../../services/dhis/pullers';
-import { DhisTranslator } from '../../../../services/dhis/DhisTranslator';
+import { DhisTranslator } from '../../../../services/dhis/translators/DhisTranslator';
 
 describe('DeprecatedEventsPuller', () => {
   let deprecatedEventsPuller;
